Refactor and finish the feature
